requirements.txt
cffi==1.14.0
chardet==3.0.4
cryptography==3.0
Django==2.2.6
django-cors-headers==3.1.1
djangorestframework==3.10.3
pdfminer.six==20200720
pycparser==2.20
pytz==2020.1
six==1.15.0
sortedcontainers==2.2.2
sqlparse==0.3.1
